Restructure C++ code
